chore(react): moved types into its own file; cleanup
